// src/queries/types.ts
export interface Character {
  id: number
  created: string
  episode: string[]
  gender: string
  image: string
  location: {
    name: string
    url: string
  }
  name: string
  origin: {
    name: string
    url: string
  }
  species: string
  status: string
  type: string
  url: string
}

export interface Episode {
  id: number
  name: string
  air_date: string
  episode: string
}